requirements.txt
connexion == 2.6.0
python_dateutil == 2.6.0
setuptools >= 21.0.0
openapi==1.1.0
psycopg2-binary
